autoops/backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# ── AWS DynamoDB ──────────────────────────────────────────────────────────────
boto3==1.34.34

# ── Auth (AlphaCore only) ─────────────────────────────────────────────────────
# django-allauth==0.61.1  # uncomment in alphacore only

# ── Task queue ────────────────────────────────────────────────────────────────
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.1

# ── Document processing ───────────────────────────────────────────────────────
pypdf2==3.0.1
python-docx==1.1.0
Pillow==10.2.0

# ── Email ─────────────────────────────────────────────────────────────────────
resend==0.8.0
requests==2.31.0
